moved FontLibrary tests to separate file
